complex area and communications data integrated v0.1(not final)
